decrement footprint
