--- src/main/java/com/alibaba/fastjson/parser/deserializer/JSONArrayDeserializer.java
@@ -1,22 +1,0 @@
-package com.alibaba.fastjson.parser.deserializer;
-
-import java.lang.reflect.Type;
-
-import com.alibaba.fastjson.JSONArray;
-import com.alibaba.fastjson.parser.DefaultJSONParser;
-import com.alibaba.fastjson.parser.JSONToken;
-
-public class JSONArrayDeserializer implements ObjectDeserializer {
-    public final static JSONArrayDeserializer instance = new JSONArrayDeserializer();
-
-    @SuppressWarnings("unchecked")
-    public <T> T deserialze(DefaultJSONParser parser, Type clazz, Object fieldName) {
-        JSONArray array = new JSONArray();
-        parser.parseArray(array);
-        return (T) array;
-    }
-
-    public int getFastMatchToken() {
-        return JSONToken.LBRACKET;
-    }
-}
