--- src/main/java/com/alibaba/fastjson/serializer/DateFormatCodec.java
@@ -1,82 +1,0 @@
-/*
- * Copyright 1999-2101 Alibaba Group.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.alibaba.fastjson.serializer;
-
-import java.io.IOException;
-import java.lang.reflect.Type;
-import java.text.SimpleDateFormat;
-
-import com.alibaba.fastjson.JSON;
-import com.alibaba.fastjson.JSONException;
-import com.alibaba.fastjson.parser.DefaultJSONParser;
-import com.alibaba.fastjson.parser.JSONToken;
-import com.alibaba.fastjson.parser.deserializer.AbstractDateDeserializer;
-import com.alibaba.fastjson.parser.deserializer.ObjectDeserializer;
-
-/**
- * @author wenshao[szujobs@hotmail.com]
- */
-public class DateFormatCodec extends AbstractDateDeserializer implements ObjectDeserializer, ObjectSerializer {
-
-    public final static DateFormatCodec instance = new DateFormatCodec();
-
-    public void write(JSONSerializer serializer, Object object, Object fieldName, Type fieldType, int features) throws IOException {
-        SerializeWriter out = serializer.out;
-
-        if (object == null) {
-            out.writeNull();
-            return;
-        }
-        
-        String pattern = ((SimpleDateFormat) object).toPattern();
-
-        if (out.isEnabled(SerializerFeature.WriteClassName)) {
-            if (object.getClass() != fieldType) {
-                out.write('{');
-                out.writeFieldName(JSON.DEFAULT_TYPE_KEY);
-                serializer.write(object.getClass().getName());
-                out.writeFieldValue(',', "val", pattern);
-                out.write('}');
-                return;
-            }
-        }
-        
-        out.writeString(pattern);
-    }
-    
-    @SuppressWarnings("unchecked")
-    protected <T> T cast(DefaultJSONParser parser, Type clazz, Object fieldName, Object val) {
-        
-        if (val == null) {
-            return null;
-        }
-
-        if (val instanceof String) {
-            String strVal = (String) val;
-            if (strVal.length() == 0) {
-                return null;
-            }
-            
-            return (T) new SimpleDateFormat(strVal);
-        }
-
-        throw new JSONException("parse error");
-    }
-
-    public int getFastMatchToken() {
-        return JSONToken.LITERAL_STRING;
-    }
-}
