+++ src/test/java/com/alibaba/json/bvt/parser/deser/generic/GenericArrayTest.java
@@ -1,20 +1,18 @@
-package com.alibaba.json.bvt.parser.deser;
+package com.alibaba.json.bvt.parser.deser.generic;
 
 import org.junit.Assert;
 
 import com.alibaba.fastjson.JSON;
-import com.alibaba.fastjson.TypeReference;
 
 import junit.framework.TestCase;
 
-public class GenericArrayTest3 extends TestCase {
+public class GenericArrayTest extends TestCase {
     
     public void test_generic() throws Exception {
         VO vo = new VO();
-        vo.values = new Pair[] {null, null};
+        vo.values = new Number[] {1, 1};
         
         String text = JSON.toJSONString(vo);
-//        VO vo1 = JSON.parseObject(text, new TypeReference<VO<Number, String>>(){} );
         VO vo1 = JSON.parseObject(text, VO.class);
         Assert.assertNotNull(vo1.values);
         Assert.assertEquals(2, vo1.values.length);
@@ -22,17 +20,12 @@ public class GenericArrayTest3 extends TestCase {
 //        Assert.assertEquals("b", vo1.values[1]);
     }
 
-    public static class A<T extends Number, S> {
-        public Pair<T, S>[] values;
+    public static class A<T extends Number> {
+        public T[] values;
     }
     
-    public static class VO extends A<Number, String> {
+    public static class VO extends A {
         
-    }
-    
-    public static class Pair<A, B> {
-        public A a;
-        public B b;
     }
 }
