+++ src/test/java/com/alibaba/json/bvt/parser/deser/generic/GenericTest.java
@@ -1,4 +1,4 @@
-package com.alibaba.json.bvt.parser.deser;
+package com.alibaba.json.bvt.parser.deser.generic;
 
 import com.alibaba.fastjson.JSON;
