+++ src/test/java/com/alibaba/json/bvt/parser/deser/generic/GenericTest2.java
@@ -1,4 +1,4 @@
-package com.alibaba.json.bvt.parser;
+package com.alibaba.json.bvt.parser.deser.generic;
 
 import java.util.List;
